Worked on the roles and the permissions

+++ src/controllers/admin-controller.ts
@@ -1,3 +1,4 @@
+import { Rolekey } from "../constants/roles";
 import { prisma } from "../lib/prisma";
 import AppError from "../utils/app-error";
 import { catchAsync } from "../utils/catch-async";
@@ -6,7 +7,14 @@ import { sendSuccess } from "../utils/response";
 export const createAdmin = catchAsync(async (req, res, next) => {
   const admin = await prisma.admin.create({
     data: {
-      userId: req.body.userId,
+      userId: req.body.protectedObject?.id,
+    },
+  });
+
+  await prisma.userRole.create({
+    data: {
+      user: { connect: { id: req.body.protectedObject.id } },
+      role: { connect: { name: Rolekey.ADMIN } },
     },
   });
 
@@ -62,9 +70,25 @@ export const updateAdmin = catchAsync(async (req, res, next) => {
 });
 
 export const deleteAdmin = catchAsync(async (req, res, next) => {
+
+  if(!req.body.protectedObject?.admin?.id) {
+    return next(new AppError("You do not have an admin account", 400));
+  }
+
   await prisma.admin.delete({
     where: {
-      id: req.params.id,
+      id: req.body.protectedObject?.admin?.id,
+    },
+  });
+
+  await prisma.userRole.delete({
+    where: {
+      userId_roleId: {
+        roleId: req.body.protectedObject?.roles?.find(
+          (role: { role: { name: string } }) => role.role.name === Rolekey.ADMIN
+        )?.roleId,
+        userId: req.body.protectedObject.id,
+      },
     },
   });
